refatoração do arquivo script.js
