+++ static/js/index.js
@@ -1,0 +1,5 @@
+export * from './utils.js';
+export * from './arquivo.js';
+export * from './solver.js';
+export * from './docenteTabela.js';
+export * from './dadosSolucao.js';
