+++ templates/index.html
@@ -235,7 +235,7 @@
     </div>
   </div>
 
-  <script type="text/javascript" src="static/script.js"></script>
+  <script type="module" src="static/js/script.js"></script>
 </body>
 
 </html>
